modify mysql engine pool size to parameter. see http://rb.corp.taobao.com/r/104655/
git-svn-id: http://code.taobao.org/svn/tfs@3279 b722c2b5-6a69-4643-bcd1-1a75825fa998

+++ branches/stable-2.5.3/src/common/internal.h
@@ -738,7 +738,7 @@ namespace tfs
       int deserialize(const char* data, const int64_t data_len, int64_t& pos);
       int serialize(char* data, const int64_t data_len, int64_t& pos) const;
       int64_t length() const;
-      uint32_t block_id_;       // block id
+      uint64_t block_id_;       // block id
       uint64_t file_id_;        // file id
       int64_t offset_;          // offset in current file
       int32_t size_;            // size of segment
